MC {splus2R}	R Documentation

Make Closure for functions

Description

MC makes closures for defining functions in a function.

Usage

MC(f, env=NULL)

Arguments

f
function

env
a list containing functions to be used in f

Details

MC declares functions to be used in f. When f is defined inside of a function, say fun, it cannot call other functions defined in fun. MC can enclose the functions needed by f and make it possible for f to call other functions defined in fun.

Author(s)

Knut M. Wittkowski kmw@rockefeller.edu

Examples

f1 <- function(x, y) x+y
f2 <- MC(function(x, y) x*y, list(f1=f1))

[Package splus2R version 1.3-5 Index]