as.rectangular {splus2R}	R Documentation

Covnert to rectangular object

Description

Rectangular data objects include matrices, data frames, and atomic vectors.

Arguments

x
object to be converted to rectangular data.

i
first (row) subscript

j
second (column) subscript.

Value

as.rectangular(x) returns x if x is already rectangular, or as.data.frame(x) if it is not.

See Also

is.rectangular.

Examples

## 
as.rectangular(list(a=1:10, b=11:20))

[Package splus2R version 1.3-5 Index]