deparseText {splus2R}	R Documentation

Deparses input argument

Description

Deparse the argument into a single string, with at most maxchars characters. New lines are turned into blanks, and truncated results end in "....".

Usage

deparseText(expr, maxchars=30)

Arguments

expr
any expression.

maxchars
maximum number of characters returned

Value

deparsed character string.

Examples

deparseText(args(lm), maxchars=20)

[Package splus2R version 1.3-5 Index]