is.inf {splus2R}	R Documentation

Infinite

Description

is.inf returns a vector of the same length as the input object, indicating which elements are infinite (not missing).

Usage

  is.inf(x)

Arguments

x
object to be tested

Details

This calls is.infinite.

This returns a vector of the same length as x; the jth element is TRUE if x[j] is infinite (i.e., equal to one of Inf or -Inf). This will be FALSE if x is not numeric or complex. Complex numbers are infinite if either the real and imaginary part is.

See Also

is.infinite

Examples

is.inf(Inf)
# [1] TRUE
is.inf(NA)
# [1] FALSE
is.inf(1)
# [1] FALSE

[Package splus2R version 1.3-5 Index]