is.missing {splus2R}	R Documentation

Check for missing values

Description

Check to see whether the input is either NA or a vector of length 0.

Usage

is.missing(x)

Arguments

x
object to check.

Value

TRUE if the input is a vector of length 0; is.na(x) otherwise.

See Also

anyMissing.

Examples

is.missing(numeric(0))
is.missing(NA)
is.missing(c(1,2,3,NA,5))

[Package splus2R version 1.3-5 Index]