is.orderable {splus2R}	R Documentation

If a value can be ordered

Description

is.orderable(x) returns !is.na().

Usage

is.orderable(x)

Arguments

x
object to be tested.

Details

x should not be a list; in that case the behavior currently differs between S-PLUS and R.

Value

is.orderable returns a logical vector of the same length as x.

See Also

is.na

Examples

x <- c(1, 4, NA, 0, 5)
is.orderable(x)
# [1]  TRUE  TRUE FALSE  TRUE  TRUE

[Package splus2R version 1.3-5 Index]