nDotArgs {splus2R}	R Documentation

Determine Number of Arguments to Function

Description

count the number of ... arguments passed.

Usage

nDotArgs(...)

Arguments

...
... arguments or real arguments in the call to the function which calls nDotArgs.

Value

the number of ... arguments in the call to the function which calls nDotArgs.

See Also

nargs.

Examples

myfun <- function(..., a=4) nDotArgs(...)
myfun()                 ## returns 0 
myfun(1:3,"bear")       ## returns 2 
myfun(a=5, 1:3, "bear") ## returns 2 (excludes a)

[Package splus2R version 1.3-5 Index]