numCols {splus2R}	R Documentation

Object dimensions

Description

Returns number of rows or number of columns of rectangular input object.

Usage

numCols(x)
numRows(x)

Arguments

x
rectangular object.

Value

the number of rows or columns of the input object.

See Also

as.data.frame,matrix,Subscript,nrow,dimnames.

Examples

x <- matrix(1:12, nrow=3)
numCols(x)
numRows(x)

[Package splus2R version 1.3-5 Index]