numericSequence {splus2R}	R Documentation

Constructor for numericSequence Class

Description

Constructor function for numericSequence objects. At least three of the four arguments must be supplied, unless the function is called with no arguments.

Usage

numericSequence(from, to, by, length.)

Arguments

from
start of the sequence.

to
end of the sequence.

by
increment for the sequence.

length.
length of the sequence, a non-negative integer.

Value

a numericSequence object with properties given by the arguments, or the default numericSequence if no arguments are supplied.

S3 METHODS

S4

supported S4 methods include: Math, Math2, Ops, Summary, [, [[, duplicated, is.na, length, match, mean, median, quantile, show, sort, summary, unique, which.na. There are also [ and [[ S4 style replacment methods available.

as

s4 style conversion methods ala as(x, foo) where foo is one of the following conversion classes: "character", "integer", "numeric".

See Also

seriesData, asSeriesData, seriesDataNew, seriesDataValid, signalSeries, as.data.frame.signalSeries, as.matrix.signalSeries, cumsum.signalSeries, deltat.signalSeries, diff.signalSeries, plot.signalSeries.

[Package splus2R version 1.3-5 Index]