signalSeries {splus2R}	R Documentation

Constructor function for the signalSeries class

Description

Construct a signalSeries object from positions and data, or return an empty signalSeries object.

Usage

signalSeries(data, positions., units, units.position, from=1, by=1)

Arguments

by
amount to skip for positions.

data
variable data, which will be converted to a rectangular object with the as.rectangular function.

from
starting value of positions.

positions.
numeric or numeric sequence object to use as the time/position values.

units
units for variable data.

units.position
units for positions.

Details

If no arguments are supplied, the default (empty) signalSeries object is returned. Otherwise, a signalSeries object is created with the given positions and data, and units if they are supplied. As an alternative to supplying the positions directly, they can be supplied by giving from and by, in which case the positions are generated as a numeric sequence with the right length to match the data.

Value

a signalSeries object with the given data and positions.

S3 METHODS

as

s4 style conversion to another class ala as(x, foo) where foo is any of the following: "character", "complex", "data.frame", "integer", "logical", "matrix", "numeric", "vector".

as.data.frame

convert to a data.frame.

as.matrix

convert to a matrix.

cumsum

cumulative summation over series.

deltat

samlig intervals of series.

diff

differencing operation applied to the series. Usage: diff(x, ...) where the ... are additional arguments sent directly to the diff function.

plot

plots the series.

See Also

numericSequence.

[Package splus2R version 1.3-5 Index]