stdev {splus2R}	R Documentation

Standard deviation

Description

Calculates the standard deviation of a series.

Usage

stdev(x, ...)

Arguments

x
input series.

...
optional arguments sent directly to the colVars function. You can control for example the removal of NA values prior to analysis via the na.rm argument, and whether or not an unbiased estimate is returned via the unbiased argument.

Value

the standard deviation of the input series.

See Also

var.

Examples

stdev(c(pi, 1, 3))

[Package splus2R version 1.3-5 Index]