subscript2d {splus2R}	R Documentation

Uniform Rectangular Data Subscripting Function

Description

subscript2d is for subscripting matrices and data frames.

Usage

subscript2d(x,i,j)
subscript2dMatrix(x,i,j)
subscript2dDataFrame(x,i,j)

Arguments

x
a matrix or data frame

i
first (row) subscript.

j
second (column) subscript.

Value

subscript2d(x,i,j) is like x[i,j,drop=F], except that it allows x[,1] (for example) for atomic vectors as well, and it always returns an object of the same class as x (that is, it does not support a drop argument).

See Also

as.data.frame, matrix.

Examples

x <- 1:10
subscript2d(x,3,1)
subscript2d(data.frame(x),3,1)
subscript2d(matrix(x),3,1)

[Package splus2R version 1.3-5 Index]